version 1

+++ backend/controllers/tareasControllers.js
@@ -3,7 +3,7 @@ const Tarea = require('../models/tareasModels')
 
 const getTareas = asyncHandler( async (req,res)=> {
    
-   const tareas = await Tarea.find({})
+   const tareas = await Tarea.find({user:req.user.id})
     res.status(200).json(tareas)
 })
 
@@ -15,8 +15,8 @@ const createTareas =asyncHandler( async(req,res)=> {
     }
 
     const tarea = await Tarea.create({
-        descripcion: req.body.descripcion
-
+        descripcion: req.body.descripcion,
+        user: req.user.id
     })
 
     res.status(201).json(tarea)
@@ -32,9 +32,18 @@ const updateTareas = asyncHandler(async(req,res)=> {
         throw new Error('No se encontro la tarea')
     }
 
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new:true})
+    //nos aseguramos que solo el dueño de la tarea  la pueda modificar
+    if(tarea.user.toString()!== req.user.id){
+        res.status(401)
+        throw new Error('No tienes permisos para modificar esta tarea')
+    }
+    else{
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new:true})
 
     res.status(200).json(tareaUpdated)
+    }
+
+    
 })
 
 const deleteTareas = asyncHandler(async(req,res)=> {
@@ -46,9 +55,17 @@ const deleteTareas = asyncHandler(async(req,res)=> {
         throw new Error('No se encontro la tarea')
     }
 
+    //nos aseguramos que solo el dueño de la tarea  la pueda modificar
+    
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('No tienes permisos para eliminar esta tarea')
+    }
+    else{
     await Tarea.deleteOne(tarea)
-
+    
     res.status(200).json({id: req.params.id})
+}
 })
 
 module.exports = {
